+++ src/main/java/com/admoliveira/exchangerate/service/RateLimiterServiceConfigProperties.java
@@ -5,7 +5,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import java.time.Duration;
 
 @ConfigurationProperties(prefix = "exchange-rate.rate-limiter-service")
-public record RateLimiterServiceConfigProperties(Duration window, int maxRequestsPerWindow) {
+public record RateLimiterServiceConfigProperties(Duration window, long maxRequestsPerWindow) {
 
     public RateLimiterServiceConfigProperties {
         if (window.toSeconds() < 1) {
